requirements.txt
streamlit==1.24.1
google-generativeai==0.1.0  
